# backend/src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Game Marketplace API",
    "version": "1.0.0",
    "description": "RAWG-based game marketplace for Buy/Sell of keys/accounts."
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1"
    }
  ],
  "paths": {
...
    "/games/{id}": {
      "get": {
        "tags": ["Games"],
        "summary": "Get game details from RAWG",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "RAWG game ID or slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Game details",
            "content": { "application/json": {} }
          }
        }
      }
    },
...
  },
...
}